# docs/swagger.yaml
basePath: /
definitions:
...
  emplacc-api_internal_dto_request.BoardCreateRequest:
    properties:
      description:
        type: string
      name:
        type: string
      project_id:
        type: string
    type: object
  emplacc-api_internal_dto_request.BoardUpdateRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  emplacc-api_internal_dto_request.UpdateProjectRequest:
    properties:
      description:
        type: string
      gitlab_project_id:
        type: integer
      gitlab_url:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
...
  emplacc-api_internal_dto_response.BoardForProjectResponse:
    properties:
      boards:
        items:
          $ref: '#/definitions/emplacc-api_internal_dto_response.BoardResponse'
        type: array
      project_id:
        type: string
    type: object
  emplacc-api_internal_dto_response.BoardListResponse:
    properties:
      boards:
        items:
          $ref: '#/definitions/emplacc-api_internal_dto_response.BoardResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
    type: object
...
  emplacc-api_internal_dto_response.BoardResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      statuses:
        items:
          $ref: '#/definitions/emplacc-api_internal_dto_response.StatusResponse'
        type: array
      updated_at:
        type: string
    type: object
  emplacc-api_internal_dto_response.BoardUniversalResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
...
  emplacc-api_internal_dto_response.ProjectResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      gitlab_project_id:
        type: integer
      gitlab_url:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  emplacc-api_internal_dto_response.ProjectUniversalResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: API для Emplacc.
  title: Emplacc API
  version: "1.0"
paths:
...
  /boards:
    post:
      consumes:
      - application/json
      description: Создает новую доску с указанными параметрами
      parameters:
      - description: Данные для создания доски
        in: body
        name: board
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_internal_dto_request.BoardCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Доска успешно создана
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.BoardUniversalResponse'
        "400":
          description: Ошибка в запросе или некорректный идентификатор проекта
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при создании доски
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создание новой доски
      tags:
      - Boards
  /boards/{id}:
    delete:
      consumes:
      - application/json
      description: Логическое удаление доски по ID (поле deleted = true), а также
        всех её статусов и задач
      parameters:
      - description: ID доски
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Доска успешно удалена
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.BoardUniversalResponse'
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Доска не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при удалении доски
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление доски
      tags:
      - Boards
    get:
      consumes:
      - application/json
      description: Получает данные доски по её уникальному идентификатору
      parameters:
      - description: ID доски
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Доска успешно получена
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.BoardResponse'
        "400":
          description: Некорректный идентификатор
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Доска не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при получении доски
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение доски по ID
      tags:
      - Boards
    patch:
      consumes:
      - application/json
      description: Обновляет данные доски по её ID
      parameters:
      - description: ID доски
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления доски
        in: body
        name: board
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_internal_dto_request.BoardUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Доска успешно обновлена
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.BoardUniversalResponse'
        "400":
          description: Ошибка в запросе или нет полей для обновления
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Доска не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при обновлении доски
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление доски
      tags:
      - Boards
  /boards/all/{page}/{pagesize}:
    get:
      consumes:
      - application/json
      description: Получает список всех досок с учетом пагинации, исключая удаленные.
      parameters:
      - description: Номер страницы
        in: path
        name: page
        required: true
        type: integer
      - description: Размер страницы
        in: path
        name: pagesize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список досок успешно получен
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.BoardListResponse'
        "400":
          description: Ошибка в запросе
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при получении досок
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение списка всех досок
      tags:
      - Boards
  /boards/project/{projectId}:
    get:
      consumes:
      - application/json
      description: Получает список досок, связанных с указанным проектом
      parameters:
      - description: ID проекта
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список досок успешно получен
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.BoardForProjectResponse'
        "400":
          description: Некорректный идентификатор проекта
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при получении досок
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение досок по ID проекта
      tags:
      - Boards
  /boards/project/{projectId}/export/xlsx:
    get:
      consumes:
      - application/json
      description: Генерирует XLSX-файл с задачами проекта, сгруппированными по доскам
        и статусам
      parameters:
      - description: ID проекта
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: XLSX-файл с задачами проекта
          schema:
            type: string
        "400":
          description: Некорректный идентификатор проекта
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при генерации отчёта
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Экспорт задач проекта в XLSX
      tags:
      - Boards
...
  /project/{id}:
    delete:
      consumes:
      - application/json
      description: Логическое удаление проекта по ID, включая все связанные доски,
        статусы и задачи (поле deleted = true)
      parameters:
      - description: ID проекта
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Проект успешно удален
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.ProjectUniversalResponse'
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Проект не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при удалении проекта
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление проекта
      tags:
      - Projects
    get:
      consumes:
      - application/json
      description: Получает данные проекта по его уникальному идентификатору
      parameters:
      - description: ID проекта
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Проект успешно получен
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.ProjectResponse'
        "400":
          description: Некорректный идентификатор проекта
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Проект не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при получении проекта
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение проекта по ID
      tags:
      - Projects
    patch:
      consumes:
      - application/json
      description: Обновляет данные проекта по его ID
      parameters:
      - description: ID проекта
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления проекта
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_internal_dto_request.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Проект успешно обновлен
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.ProjectUniversalResponse'
        "400":
          description: Некорректный идентификатор или ошибка в запросе
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при обновлении проекта
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление проекта
      tags:
      - Projects
...
  /status/board/{board_id}:
    get:
      description: Получение списка статусов, связанных с доской по её ID, включая
        все задачи для каждого статуса
      parameters:
      - description: ID доски
        in: path
        name: board_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список статусов для доски
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.StatusByBoardIdResponse'
        "400":
          description: Ошибка при парсинге ID доски
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при получении статусов
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение статусов по ID доски
      tags:
      - Statuses
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"